refactor(task-engine): extract task-runner leaf helpers into task-engine [SHR-187]
Move dependency-free leaf helpers out of server/task-runner.ts into
server/task-engine/{sessions,reconcile,git,launch}.ts; task-runner.ts
re-exports them so all callers and the 2742-line test are unchanged.
Identical child_process/execTmux/fs transport and git/tmux argv.

Modular 04 step 04.1.
